@@ -42,60 +44,70 @@
     "moment": "^2.27.0",
     "normalize.css": "^4.2.0",
     "normalizr": "^3.6.0",
+    "notistack": "^1.0.1",
     "path-to-regexp": "^2.0.0",
     "prop-types": "^15.6.0",
     "query-string": "^5.0.0",
-    "react": "^16.8.0",
+    "react": "^16.13.1",
     "react-big-calendar": "^0.26.0",
     "react-dom": "^16.8.0",
     "react-helmet": "^5.2.0",
     "react-highlight": "github:briancappello/react-highlight#react-v16-compiled",
+    "react-intl": "^5.8.8",
     "react-leaflet-control": "^2.1.2",
     "react-loadable": "^4.0.5",
-    "react-redux": "^5.0.6",
+    "react-redux": "^7.2.2",
     "react-redux-loading-bar": "^2.9.2",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^5.0.0-alpha.6",
+    "react-router": "5.2.0",
+    "react-router-dom": "5.2.0",
     "react-slick": "^0.26.1",
     "react-svg": "^2.1.22",
     "redux": "^3.7.2",
-    "redux-form": "^7.1.0",
+    "redux-form": "^8.3.6",
+    "redux-orm": "^0.16.2",
+    "redux-promising-modals": "^2.1.2",
     "redux-saga": "^0.15.6",
     "reselect": "^4.0.0",
     "super-skeleton": "github:briancappello/super-skeleton"
   },
   "devDependencies": {
     "uglifyjs-webpack-plugin": "1.0.1",
+    "@babel/core": "^7.12.3",
+    "@babel/plugin-proposal-export-default-from": "^7.12.1",
+    "@babel/plugin-proposal-function-bind": "^7.12.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
+    "@babel/runtime": "^7.12.1",
+    "@formatjs/cli": "^2.13.5",
     "add-asset-html-webpack-plugin": "^2.1.1",
-    "babel-core": "^6.26.0",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.2",
+    "babel-jest": "^26.6.1",
+    "babel-loader": "^8.1.0",
     "babel-plugin-dynamic-import-node": "^1.0.2",
+    "babel-plugin-react-intl": "^8.2.10",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-plugin-transform-react-display-name": "^6.25.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
     "css-loader": "^0.28.5",
     "express": "^4.15.4",
     "express-http-proxy": "^1.6.2",
+    "extract-react-intl-messages": "^4.1.1",
     "file-loader": "^0.11.2",
     "html-webpack-plugin": "^2.30.1",
     "ip": "^1.1.5",
-    "jest-cli": "^20.0.4",
+    "jest-cli": "^26.6.1",
     "json-loader": "^0.5.7",
     "node-sass": "^4.5.3",
-    "react-hot-loader": "^3.0.0-beta.7",
+    "react-hot-loader": "^4.13.0",
     "react-leaflet": "^2.7.0",
     "redbox-react": "^1.5.0",
     "resolve-url-loader": "^2.1.0",
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.6",
-    "shelljs": "^0.7.8",
+    "shelljs": "^0.8.4",
     "style-loader": "^0.18.2",
     "url-loader": "^0.6.2",
     "webpack": "^3.6.0",
