@@ -4,7 +4,7 @@
 # Common Python library imports
 from sqlalchemy.orm.collections import attribute_mapped_collection
 import sqlalchemy as sa
-import enum
+from sqlalchemy.orm import backref
 
 # Pip package imports
 from geoalchemy2 import Geometry
@@ -25,20 +25,6 @@ from backend.database import (
 )
 from .reference_parcel_relation import ReferenceParcelRelation
 from .reference_parcel_mixin import ReferenceParcelMixin
-from .reference_parcel_property import ReferenceParcelProperty
-
-def create_production(production):
-    from ..models import ReferenceParcelProduction
-    return ReferenceParcelProduction(production=production)
-
-
-class ReferenceParcelTypes(enum.Enum):
-    AgriculturalParcel = 'AgriculturalParcel'
-    CadastralParcel = 'CadastralParcel'
-    FarmersBlock = 'FarmersBlock'
-    PhysicalBlock = 'PhysicalBlock'
-
-
 
 class ReferenceParcel(ReferenceParcelMixin, TimestampMixin, BaseModel):
     #id = Column(BigInteger, primary_key=True, autoincrement=True)
@@ -48,49 +34,45 @@ class ReferenceParcel(ReferenceParcelMixin, TimestampMixin, BaseModel):
     geometry = Column(Geometry("POLYGON"))  # TODO: Do I need to specify the srid or not?
 
     total_area = Column(Float(), nullable=False)
-    eligible_area = Column(Float(), nullable=False)
 
-    soil_type_id = foreign_key('SoilType', nullable=False)
-    soil_type = relationship('SoilType', uselist=False)
+    # Ancestor - Descendant
+    # TODO: Fix this
+    ancestor_id = foreign_key('ReferenceParcel', fk_col='parcel_id', nullable=True)
+    ancestor = relationship("ReferenceParcel",
+                            backref='descendant',
+                            remote_side='ReferenceParcel.parcel_id',
+                            primaryjoin=('reference_parcel.c.parcel_id==reference_parcel.c.ancestor_id')
+                            )
 
-    # Reference Parcel Type relationship definition
-    agricultural_type_id = foreign_key('AgriculturalType', nullable=False)
-    agricultural_type = relationship('AgriculturalType', uselist=False)
 
     # Season relationship definition
-    # FIXME: Would be better to set nullable=False, but that require a major update in the unit test setup.
     season_id = foreign_key('Season', nullable=True, onupdate="CASCADE", ondelete="CASCADE")
-    season = relationship('Season', back_populates='reference_parcels')
+    season = relationship('Season', back_populates='parcels')
+
+    # Keep relationship definition in the reference parcel.
+    # TODO: This is perfect!!! Everyting working as expected, DON'T TOUCH IT!!
+    block_parcels = relationship('ReferenceParcelRelation',
+                         primaryjoin="ReferenceParcel.parcel_id==reference_parcel_relation.c.parcel_id",
+                         backref='block',
+                         cascade="all, delete")
+
+    parcel_blocks = relationship('ReferenceParcelRelation',
+                          primaryjoin="ReferenceParcel.parcel_id==reference_parcel_relation.c.block_id",
+                          backref='parcel',
+                          cascade="all, delete")
 
     # Group relationship definition
-    groups = association_proxy('group_parcels', 'group',)
+    blocks = association_proxy('block_parcels', 'block')
 
     # Parcel relationship definition
-    parcels_add = association_proxy('parcel_groups', 'parcel',)
+    parcels_add = association_proxy('parcel_blocks', 'parcel')
 
     @property
     def parcels(self):
         return ReferenceParcel.join(ReferenceParcelRelation, (ReferenceParcelRelation.parcel_id == ReferenceParcel.parcel_id)).filter(
-            ReferenceParcelRelation.group_id == self.parcel_id).all()
+            ReferenceParcelRelation.block_id == self.parcel_id).all()
         #return session.query(Node).join(NodeRelation, (NodeRelation.child_id == Node.id)).filter(
             #NodeRelation.parent_id == self.id).all()
-
-    # Production relationship definition
-    reference_parcel_productions = relationship('ReferenceParcelProduction', back_populates='reference_parcel',
-                                            cascade='all, delete-orphan')
-    productions = association_proxy('reference_parcel_productions', 'production',
-                                      creator=lambda production: create_production(production))
-
-
-    property = relationship(
-        "ReferenceParcelProperty", collection_class=attribute_mapped_collection("key")
-    )
-
-    properties = association_proxy(
-        "property",
-        "value",
-        creator=lambda key, value: ReferenceParcelProperty(key=key, value=value),
-    )
 
     __repr_props__ = ('id', 'title', 'parcels')
 
