@@ -4,6 +4,7 @@
 # Common Python library imports
 # Pip package imports
 import sqlalchemy as sa
+from sqlalchemy.orm import backref
 
 # Internal package imports
 from backend.database import (
@@ -20,30 +21,15 @@ from backend.database import (
 class ReferenceParcelRelation(BaseModel, TimestampMixin):
     """Join table between User and Role"""
 
-    group_id = foreign_key('ReferenceParcel', fk_col='parcel_id', primary_key=True, onupdate="CASCADE", ondelete="CASCADE")
-    parcel_id = foreign_key('ReferenceParcel', fk_col='parcel_id', primary_key=True, onupdate="CASCADE", ondelete="CASCADE")
-    #group_id = foreign_key('ReferenceParcel', primary_key=True)
-    #parcel_id = foreign_key('ReferenceParcel', primary_key=True)
+    block_id = foreign_key('ReferenceParcel', fk_col='parcel_id', primary_key=True, onupdate="CASCADE", ondelete="CASCADE")
+    parcel_id = foreign_key('ReferenceParcel', fk_col='parcel_id', primary_key=True, onupdate="CASCADE")
 
-    group = relationship('ReferenceParcel',
-                         #primaryjoin="ReferenceParcelRelation.group_id == reference_parcel.c.id",
-                         backref='group_parcels',
-                         foreign_keys=parcel_id,
-                         cascade="all, delete"
-                         )
+    __repr_props__ = ('block_id', 'parcel_id')
 
-    parcel = relationship('ReferenceParcel',
-                          #primaryjoin="ReferenceParcelRelation.parcel_id == reference_parcel.c.id",
-                          backref='parcel_groups',
-                          foreign_keys=group_id,
-                          cascade="all, delete"
-                          )
-
-    __repr_props__ = ('group_id', 'parcel_id')
-
-    def __init__(self, group=None, parcel=None, **kwargs):
+    def __init__(self, block=None, parcel=None, **kwargs):
+        print("Adding relation - Block: %s Parcel: %s" % (block, parcel))
         super().__init__(**kwargs)
         if parcel:
             self.parcel = parcel
-        if group:
-            self.group = group
+        if block:
+            self.block = block
