@@ -17,6 +17,7 @@ from backend.database import (
 )
 
 class Unit(Model):
-    title = Column(String(64), nullable=False)
+    title = Column(String(64))
+    so_unit = Column(String(16), nullable=True)
 
     __repr_props__ = ('id', 'title')
