@@ -8,8 +8,8 @@ from .unit import UnitSerializer, UnitListSerializer
 from .soil_type import SoilTypeSerializer, SoilTypeListSerializer
 from .country import CountrySerializer, CountryListSerializer
 from .region import RegionSerializer, RegionListSerializer
-from .agricultural_product import AgriculturalProductSerializer, AgriculturalProductListSerializer
-from .agricultural_product_fact import AgriculturalProductFactSerializer, AgriculturalProductFactListSerializer
-from .specific_product import SpecificProductSerializer, SpecificProductListSerializer
+#from .agricultural_product import AgriculturalProductSerializer, AgriculturalProductListSerializer
+#from .agricultural_product_fact import AgriculturalProductFactSerializer, AgriculturalProductFactListSerializer
+#from .specific_product import SpecificProductSerializer, SpecificProductListSerializer
 
 from .agricultural_type import AgriculturalTypeSerializer, AgriculturalTypeListSerializer
