@@ -1,13 +1,14 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import { ConnectedRouter } from 'react-router-redux'
+import { ConnectedRouter } from 'connected-react-router'
+import { SnackbarProvider } from 'notistack';
 import Helmet from 'react-helmet'
 import { StylesProvider } from '@material-ui/core/styles';
 
-import { NavBar, ProgressBar, SideMenu } from 'components'
-import { SITE_NAME, COPYRIGHT } from 'config'
+import { NavBar, ProgressBar } from 'components'
+import { LanguageProvider, ModalProvider } from 'site/components'
+import { SITE_NAME } from 'config'
 import Routes from 'routes'
-
 
 const AppLayout = () => (
   <div>
@@ -15,22 +16,27 @@ const AppLayout = () => (
             defaultTitle={SITE_NAME}
     />
     <ProgressBar />
-    <main className="page-container page-panels">
-      <SideMenu />
+    <main className="app-container">
+      <ModalProvider />
+      <NavBar />
       <Routes />
     </main>
-    <footer className="center">
-       Copyright {new Date().getFullYear()} {COPYRIGHT}
-    </footer>
   </div>
 )
 
-export default (props) => (
-  <StylesProvider injectFirst>
-    <Provider store={props.store}>
-      <ConnectedRouter history={props.history}>      
-          <AppLayout />
-      </ConnectedRouter>
-    </Provider>
-  </StylesProvider>
-)
+export default (props) => {
+  console.log("App props: ", props)
+  return (
+    <StylesProvider injectFirst>
+        <Provider store={props.store}>
+          <LanguageProvider messages={props.messages}>
+            <ConnectedRouter history={props.history}>      
+              <SnackbarProvider maxSnack={3}>
+                <AppLayout />
+              </SnackbarProvider>
+            </ConnectedRouter>
+          </LanguageProvider>
+        </Provider>
+      </StylesProvider>
+  )
+}
