@@ -1,5 +1,4 @@
 import DocComponent from './DocComponent'
 import PageContent from './PageContent'
-import PageHeader from './PageHeader'
 
-export { DocComponent, PageContent, PageHeader }
+export { DocComponent, PageContent }
