@@ -3,6 +3,11 @@ export { DocComponent, PageContent, PageHeader } from './Content'
 export { Date, LongDate, Time, DateTime, LongDateTime } from './Date'
 export { Loading, ProgressBar, Spinner } from './Loading'
 export { HashLink, NavLink, NavBar, ScrollIntoView } from './Nav'
-export { SideMenu, SideBarItem } from './SideMenu'
 export { ListElementSmall, ListElementLong, ListLayout } from './List'
 export { Table } from './Table'
+export { default as Stepper } from './Stepper'
+
+export { MenuButton, NestedMenuItem, MenuItem } from './Menu'
+
+export { SubmitButton, BackButton, SplitButton } from './Button'
+export { HeaderContent, MasterDetail, Content } from './Layout'
