@@ -1,20 +1,29 @@
 import { combineReducers } from 'redux'
 import formReducer from 'redux-form/es/reducer'
-import { routerReducer } from 'react-router-redux'
+import { connectRouter } from 'connected-react-router'
 import { loadingBarReducer } from 'react-redux-loading-bar'
+import { modalsReducer } from 'redux-promising-modals';
 
 import securityReducer from 'security/reducer'
-import flashReducer from 'site/reducers/flash'
+import notificationReducer from 'site/reducers/notification'
+import localeReducer from 'site/reducers/language'
+import history from 'utils/history'
 
+import { createReducer as createReducerORM } from 'redux-orm'
+import orm from 'orm'
 
 export default function createReducer(injectedReducers) {
   return combineReducers({
     security: securityReducer,
-    flash: flashReducer,
+    notification: notificationReducer,
+    locale: localeReducer,
 
     form: formReducer,
-    routing: routerReducer,
+    router: connectRouter(history),
     loadingBar: loadingBarReducer,
+    modals: modalsReducer,
+
+    entities: createReducerORM(orm),
 
     ...injectedReducers,
   })
