@@ -6,12 +6,19 @@ import reduxForm from 'redux-form/es/reduxForm'
 import { parse } from 'query-string'
 
 import { login } from 'security/actions'
-import { DangerAlert, PageContent } from 'components'
+import Paper from '@material-ui/core/Paper';
+import Container from '@material-ui/core/Container';
+import Grid from '@material-ui/core/Grid';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+
+import { PageContent } from 'components'
 import { NavLink } from 'components/Nav'
 import { HiddenField, PasswordField, TextField } from 'components/Form'
 import { ROUTES } from 'routes'
 import { injectSagas } from 'utils/async'
 
+import './login.scss'
 
 const FORM_NAME = 'login'
 
@@ -23,36 +30,57 @@ const Login = (props) => {
       <Helmet>
         <title>Login</title>
       </Helmet>
-      <div className="row">
-        <div className="six cols offset-by-three">
-          <h1>Log in!</h1>
-          {error && <DangerAlert>{error}</DangerAlert>}
-          {isDev && <p>Hint: a@a.com / password</p>}
+      <Container className="login-outer" maxWidth={false}>
+        <Paper elevation={3}>
+          <Typography variant="h4" gutterBottom>
+            Login
+          </Typography>
           <form onSubmit={handleSubmit(login)}>
-            <HiddenField name="redirect" />
-            <TextField name="email"
-                       label="Email or Username"
-                       className="full-width"
-                       autoFocus
-            />
-            <PasswordField name="password"
-                           className="full-width"
-            />
-            <div className="row">
-              <button type="submit"
-                      className="btn btn-primary"
-                      disabled={pristine || submitting}
-              >
-                {submitting ? 'Logging in...' : 'Submit'}
-              </button>
-              <NavLink to={ROUTES.ForgotPassword}
-                       className="pull-right"
-                       style={{lineHeight: '38px'}}
-              />
-            </div>
-          </form>
-        </div>
-      </div>
+            <Grid 
+              className="login-inner"
+              container
+              spacing={3}
+            >
+                <HiddenField name="redirect" />
+                <Grid item xs={12}>
+                  <TextField name="email"
+                            formProps={{
+                              fullWidth: true
+                            }}
+                            label="Email or Username"
+                            autoFocus
+                  />
+                </Grid>
+                <Grid item xs={12}>
+                  
+                    <PasswordField name="password"
+                                  formProps={{
+                                    fullWidth: true
+                                  }}
+                                  label="Password"
+                    />
+                  <div>
+                    <NavLink to={ROUTES.ForgotPassword}
+                            className="forgot-password"
+                            style={{lineHeight: '38px'}}
+                    />
+                    </div>
+                </Grid>
+                <Grid item xs={12}>
+                  <Button 
+                    type="submit"
+                    disabled={pristine || submitting}
+                    variant="contained"
+                    color="primary"
+                  >
+                    {submitting ? 'Logging in...' : 'Submit'}
+                  </Button>
+                </Grid>
+            </Grid>
+              
+            </form>
+        </Paper>   
+      </Container>         
     </PageContent>
   )
 }
