@@ -1,6 +1,5 @@
 import { call, put, takeLatest } from 'redux-saga/effects'
 
-import { flashSuccess } from 'site/actions'
 import { createRoutineFormSaga } from 'sagas'
 
 import { changePassword } from 'security/actions'
@@ -14,7 +13,7 @@ export const changePasswordSaga = createRoutineFormSaga(
   function *successGenerator(payload) {
     const { token } = yield call(SecurityApi.changePassword, payload)
     yield put(changePassword.success({ token }))
-    yield put(flashSuccess('Your password has been successfully changed.'))
+    //yield put(flashSuccess('Your password has been successfully changed.'))
   },
 )
 
