@@ -1,7 +1,6 @@
 import { call, put, takeLatest } from 'redux-saga/effects'
-import { push } from 'react-router-redux'
+import { push } from 'connected-react-router'
 
-import { flashSuccess } from 'site/actions'
 import { createRoutineFormSaga } from 'sagas'
 
 import { login } from 'security/actions'
@@ -17,7 +16,7 @@ export const loginSaga = createRoutineFormSaga(
     const { token, user } = yield call(SecurityApi.login, payload)
     yield put(login.success({ token, user }))
     yield put(push(redirect))
-    yield put(flashSuccess('You have been successfully logged in.'))
+    //yield put(flashSuccess('You have been successfully logged in.'))
   },
 )
 
