@@ -1,7 +1,6 @@
 import { call, put, takeLatest } from 'redux-saga/effects'
-import { push } from 'react-router-redux'
+import { push } from 'connected-react-router'
 
-import { flashSuccess } from 'site/actions'
 import { ROUTES, ROUTE_MAP } from 'routes'
 import { createRoutineSaga } from 'sagas'
 
@@ -16,8 +15,8 @@ export const logoutSaga = createRoutineSaga(
   function *successGenerator() {
     yield call(SecurityApi.logout)
     yield put(logout.success())
-    yield put(push(ROUTE_MAP[ROUTES.Home].path))
-    yield put(flashSuccess('You have been successfully logged out.'))
+    yield put(push(ROUTE_MAP[ROUTES.DashboardHome].path))
+    //yield put(flashSuccess('You have been successfully logged out.'))
   },
 )
 
