@@ -1,7 +1,6 @@
 import { call, put, takeLatest } from 'redux-saga/effects'
-import { push } from 'react-router-redux'
+import { push } from 'connected-react-router'
 
-import { flashSuccess } from 'site/actions'
 import { ROUTES, ROUTE_MAP } from 'routes'
 import { createRoutineFormSaga } from 'sagas'
 
@@ -17,8 +16,8 @@ export const resetPasswordSaga = createRoutineFormSaga(
     const { token: resetToken, ...payload } = actionPayload
     const { token, user } = yield call(SecurityApi.resetPassword, resetToken, payload)
     yield put(resetPassword.success({ token, user }))
-    yield put(push(ROUTE_MAP[ROUTES.Home].path))
-    yield put(flashSuccess('Welcome back! Your password has been successfully changed.'))
+    yield put(push(ROUTE_MAP[ROUTES.DashboardHome].path))
+    //yield put(flashSuccess('Welcome back! Your password has been successfully changed.'))
   },
 )
 
