@@ -1,43 +1,46 @@
 import { createRoutine } from 'actions'
 
+export const ENQUEUE_NOTIFICATION = 'ENQUEUE_NOTIFICATION'
+export const CLOSE_NOTIFICATION = 'CLOSE_NOTIFICATION'
+export const REMOVE_NOTIFICATION = 'REMOVE_NOTIFICATION'
 
-export const FLASH_SUCCESS = 'FLASH_SUCCESS'
-export const FLASH_INFO = 'FLASH_INFO'
-export const FLASH_WARNING = 'FLASH_WARNING'
-export const FLASH_DANGER = 'FLASH_DANGER'
-export const FLASH_CLEAR = 'FLASH_CLEAR'
+export const CHANGE_LOCALE = 'CHANGE_LOCALE'
 
-export function flashSuccess(title, message) {
+export const enqueueNotification = (notification) => {
+  const key = notification.options && notification.options.key
+
   return {
-    type: FLASH_SUCCESS,
-    payload: { title, message },
+    type: ENQUEUE_NOTIFICATION,
+    payload: {
+      notification: {
+        ...notification,
+        key: key || new Date().getTime() + Math.random()
+      }
+    }
   }
 }
 
-export function flashInfo(title, message) {
-  return {
-    type: FLASH_INFO,
-    payload: { title, message },
+export const closeNotification = key => ({
+  type: CLOSE_NOTIFICATION,
+  payload: {
+    dismisAll: !key, // dismiss all if no key has been defined
+    key,
   }
-}
+})
 
-export function flashWarning(title, message) {
-  return {
-    type: FLASH_WARNING,
-    payload: { title, message },
+export const removeNotification = key => ({
+  type: REMOVE_NOTIFICATION,
+  payload: {
+    key,
   }
-}
+})
 
-export function flashDanger(title, message) {
+export function changeLocale(languageLocale) {
   return {
-    type: FLASH_DANGER,
-    payload: { title, message },
-  }
-}
-
-export function flashClear() {
-  return {
-    type: FLASH_CLEAR,
+    type: CHANGE_LOCALE,
+    payload: {
+      locale: languageLocale,
+    }
   }
 }
 
