@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import { useStore, ReactReduxContext } from 'react-redux';
 import hoistNonReactStatics from 'hoist-non-react-statics'
 import get from 'lodash/get'
 
@@ -29,9 +30,8 @@ export default (props) => (WrappedComponent) => {
 
   class InjectSaga extends React.Component {
     static WrappedComponent = WrappedComponent
-    static contextTypes = {
-      store: PropTypes.object.isRequired,
-    }
+    static contextType = ReactReduxContext;
+
     static displayName = `withSaga(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`
 
     componentWillMount() {
